requirements.txt
Flask
flask-cors
google-generativeai
# pip install -r requirements.txt
#first